requirements.txt
fastapi
uvicorn[standard]
sqlmodel
scikit-learn
numpy
Pillow
python-multipart
pydantic-settings